requirements.txt
faiss_cpu==1.8.0.post1
langchain==0.2.14
langchain_community==0.2.12
langchain_core==0.2.34
langchain_experimental==0.0.64
langchain_openai==0.1.22
pandas==2.2.2
python-dotenv==1.0.1
streamlit==1.33.0
statsmodels
flagembedding